refactor: 토큰 전처리 클래스 분리
- TicketTokenizer: 문자열 -> 숫자 리스트
- TicketGuard: 숫자 리스트 조건 검사 -> 정렬된 숫자 반환
